+++ src/main/kotlin/com/crypted/explorer/gateway/action/BlockAction.kt
@@ -55,7 +55,7 @@ class BlockAction(private val blockService: BlockService)  {
     @ApiResponse(responseCode = "501", description = "Invalid Request", content = [Content(schema = Schema(implementation = Result::class))])
     @ApiResponse(responseCode = "502", description = "Invalid Parameter", content = [Content(schema = Schema(implementation = Result::class))])
     @ApiResponse(responseCode = "504", description = "Missing parameter", content = [Content(schema = Schema(implementation = Result::class))])
-    fun getInfoByBlockNumber(@Parameter(description = "blockNumber", required = true) @PathVariable("blockNumber") @NotNull @Min(0) blockNumber: Int): Result<BlockInfoResp?> {
+    fun getInfoByBlockNumber(@Parameter(description = "blockNumber", required = true) @PathVariable("blockNumber") @NotNull @Min(0) blockNumber: Long): Result<BlockInfoResp?> {
 
         LOG.info(Log.format("success", Log.kv("api", "block/")))
